requirements.txt
gunicorn 
flask 
requests
py-openaq
python-dateutil
